+++ swarmauri_experimental/tests/integration/vector_stores/SpatialDocEmbedding_test.py
@@ -12,20 +12,6 @@ def test_ubc_resource():
     test()
 
 
-@pytest.mark.unit
-def test_ubc_type():
-    assert SpatialDocEmbedding().type == "SpatialDocEmbedding"
-
-
-@pytest.mark.unit
-def test_serialization():
-    embedder = SpatialDocEmbedding()
-    assert (
-        embedder.id
-        == SpatialDocEmbedding.model_validate_json(embedder.model_dump_json()).id
-    )
-
-
 @pytest.mark.acceptance
 def test_fit_transform():
     embedder = SpatialDocEmbedding()
